#YA QUEDO, FINAL, con permisos

+++ inc/agregar_sucursal.php
@@ -184,9 +184,25 @@ if (!empty($_POST))
                     <div class="form-group col-lg-6">
                         
                             <label for="newsucursal">Tipo</label>
-                            <button data-toggle="modal" data-target="#nuevo_tipo" title="Agregar nuevo Tipo" class="btn btn-primary btn-xs" type="button" href="#" ><i class="fas fa-plus"></i></button>
-                            <button disabled data-toggle="modal" data-target="#editar_tipo" onclick="editar_tipo();" title="Editar Tipo" class="btn btn-success btn-xs" type="button" href="#" id="btn_editartipo"><i class="fas fa-edit"></i></button>
-                            <button disabled onclick="eliminar_tipo();" title="Eliminar Tipo" class="btn btn-danger btn-xs" type="button" href="#" id="btn_eliminartipo"><i class="fas fa-trash"></i></button>
+                            <?php 
+                                if($_SESSION['rol'] == "SuperAdmin" or $_SESSION['rol'] == "Administrador")
+                                {
+                                    ?>
+                                    <button data-toggle="modal" data-target="#nuevo_tipo" title="Agregar nuevo Tipo" class="btn btn-primary btn-xs" type="button" href="#" ><i class="fas fa-plus"></i></button>
+                                    <button disabled data-toggle="modal" data-target="#editar_tipo" onclick="editar_tipo();" title="Editar Tipo" class="btn btn-success btn-xs" type="button" href="#" id="btn_editartipo"><i class="fas fa-edit"></i></button>
+                                    <button disabled onclick="eliminar_tipo();" title="Eliminar Tipo" class="btn btn-danger btn-xs" type="button" href="#" id="btn_eliminartipo"><i class="fas fa-trash"></i></button>
+                                    <?php
+                                }
+                                else
+                                {
+                                    ?>
+                                    <button disabled="disabled" title="Agregar nuevo Tipo" class="btn btn-primary btn-xs" type="button"><i class="fas fa-plus"></i></button>
+                                    <button disabled="disabled" title="Editar Tipo" class="btn btn-success btn-xs" type="button"><i class="fas fa-edit"></i></button>
+                                    <button disabled="disabled" title="Eliminar Tipo" class="btn btn-danger btn-xs" type="button"><i class="fas fa-trash"></i></button>
+                                    <?php 
+                                } 
+                             ?>
+
                             <select class="form-control" id="tipo" name="tipo">
                                 <option selected hidden>Selecciona un tipo</option>
                                 <?php
